v1 transformer uses calculateGridItemDimensions
